backend/requirements.txt
fastapi==0.68.2
uvicorn[standard]==0.15.0
supabase==1.2.0
pydantic==1.10.13
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.24.1
sqlalchemy==1.4.53
psycopg2-binary==2.9.9
